Extract streaming query orchestration into query service
